requirements.txt
fastapi
uvicorn[standard]
streamlit
plotly
pandas
scikit-learn
pyyaml
python-dotenv
groq
python-multipart
requests
